requirements.txt
# Use Python 3.11

attrs==25.3.0
cattrs==25.1.1
tqdm==4.67.1
hydra-core==1.3.2
omegaconf==2.3.0
requests==2.32.4
numpy==2.3.2
pandas==2.3.1
pyarrow==21.0.0
tabulate==0.9.0
scipy==1.16.1
scikit-learn==1.7.1
matplotlib==3.10.5
torch==2.7.1
transformers==4.54.1
sentence-transformers==5.0.0
nltk==3.9.1
evaluate==0.4.5
tiktoken==0.9.0
rank-bm25==0.2.2
gsutil==5.35